ether-service/requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
grpcio==1.25.0
grpcio-tools==1.25.0
idna==3.1
idna-ssl==1.1.0
isort==5.6.4
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
protobuf==3.14.0
PyYAML==5.1.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.6.3
PyJWT==1.7.1
pymongo==3.11.0
motor==2.2.0
aio-pika==6.7.1
kubernetes==11.0.0
requests
celery==4.2.0
kombu==4.6.11